add unit tests, make struct fields private, make input options more generic

+++ copy-folder/src/transfer.rs
@@ -9,21 +9,28 @@ use rayon::prelude::*;
 
 #[derive(Debug)]
 pub struct Transfer {
-    pub src: PathBuf,
-    pub dest: PathBuf,
+    src: PathBuf,
+    dest: PathBuf,
 }
 
 impl Transfer {
-    pub fn run(&self, files: Vec<PathBuf>) -> Result<()> {
-        let mut summary = Summary {
-            start: Instant::now(),
-            successes: 0,
-            failures: 0,
-            gb: 0.0,
-        };
+    const GB: u64 = 1_024u64.pow(3);
+
+    pub fn new(src: &Path, dest: &Path) -> Self {
+        Self {
+            src: src.to_path_buf(),
+            dest: dest.to_path_buf(),
+        }
+    }
+
+    pub fn run<I>(&self, files: I) -> Result<()>
+    where
+        I: IntoIterator<Item = PathBuf>,
+    {
+        let files: HashSet<_> = files.into_iter().collect();
 
         // create any missing parent directories so file copying is independent
-        let parents: HashSet<PathBuf> = files
+        let parents: HashSet<_> = files
             .iter()
             .map(|file| self.dest.join(file))
             .map(|path| path.parent().unwrap().to_path_buf())
@@ -36,21 +43,19 @@ impl Transfer {
         }
 
         // do the actual copying in parallel
-        let results: Vec<Result<u64>> =
-            files.into_par_iter().map(|file| self.copy(&file)).collect();
-
-        for result in results {
-            summary.add(result);
-        }
-
-        summary.done();
-
+        let start = Instant::now();
+        let results = files.into_par_iter().map(|file| self.copy(&file)).collect();
+        Self::summarize(start, results);
         Ok(())
     }
 
     fn copy(&self, file: &Path) -> Result<u64> {
         let from = self.src.join(file);
         let to = self.dest.join(file);
+        // skip file if it already exist in destination
+        if to.exists() {
+            return Ok(0);
+        }
         info!("copying: {:?} -> {:?}", from, to);
         fs::copy(&from, &to)?;
         self.cleanup(&to)?;
@@ -71,37 +76,28 @@ impl Transfer {
 
     #[cfg(not(target_os = "macos"))]
     fn cleanup(&self, _path: &Path) -> Result<()> {
+        // currently outside of macos no cleanup is needed
         Ok(())
     }
-}
 
-#[derive(Debug)]
-struct Summary {
-    start: Instant,
-    successes: usize,
-    failures: usize,
-    gb: f64,
-}
-
-impl Summary {
-    const GB: u64 = 1_024u64.pow(3);
-
-    fn add(&mut self, result: Result<u64>) {
-        match result {
-            Ok(bytes) => {
-                self.successes += 1;
-                self.gb += bytes as f64 / Self::GB as f64;
-            }
-            Err(_) => {
-                self.failures += 1;
+    fn summarize(start: Instant, results: Vec<Result<u64>>) {
+        let mut successes = 0;
+        let mut failures = 0;
+        let mut gb = 0.0;
+        for result in results {
+            match result {
+                Ok(bytes) => {
+                    successes += 1;
+                    gb += bytes as f64 / Self::GB as f64;
+                }
+                Err(_) => {
+                    failures += 1;
+                }
             }
         }
-    }
-
-    fn done(&self) {
-        info!("duration: {:.2?}", self.start.elapsed());
-        info!("successes: {}", self.successes);
-        info!("failures: {}", self.failures);
-        info!("gigabytes: {:.4}", self.gb);
+        info!("duration: {:.2?}", start.elapsed());
+        info!("successes: {}", successes);
+        info!("failures: {}", failures);
+        info!("gigabytes: {:.4}", gb);
     }
 }
